+++ ui/UI.java
@@ -1,45 +1,13 @@
 package ui;
 
 import java.io.*;
-import control.Control;
-import models.Cell;
 
 // uiパッケージのメインクラス
 public class UI {
 
+    // ゲームの終了
     protected static void exitGame() {
-        Control.exitGame();
-    }
-
-    // fieldを表示
-    protected void displayField(Cell[][] fieldmap) {
-        final int MAX_Y = fieldmap.length;
-        final int MAX_X = fieldmap[0].length;
-
-        clearScreen();
-
-        // xの目盛表示
-        for(int i = 0; i <= MAX_X; i++)
-            System.out.printf("%2d ", i);
-        System.out.println("");
-
-        // yの目盛表示
-        for(int i = 0; i < MAX_Y; i++) {
-            System.out.printf("%2d ", i + 1);
-            // fieldの表示
-            for(int j = 0; j < MAX_X; j++)
-                System.out.printf("%2s ", fieldmap[i][j].character);
-            System.out.println("");
-        }
-    }
-
-    // shellコンソール表示のクリア
-    private void clearScreen() {
-        try {
-            new ProcessBuilder("/bin/bash", "-c", "clear").inheritIO().start().waitFor();
-        } catch(Exception e) {
-            e.printStackTrace();
-            System.exit(1);
-        }
+        System.out.println("Good bye!");
+        System.exit(0);
     }
 }
